Add files via upload
This case study is on a Multinational Retail Corporation which is a globally renowned brand and a prominent retailer in the United States.
Using SQL to analyze the company extensive dataset and gain valuable insights into Company's operations in Brazil. The information can shed light on various aspects of the business, such as order processing, pricing strategies, payment and shipping efficiency, customer demographics, product characteristics, and customer satisfaction levels.